requirements.txt
pandas~=3.0.1
joblib~=1.5.3
scikit-learn~=1.8.0
sentence-transformers~=5.3.0
python-dotenv~=1.2.2
groq~=1.1.2